Split identifier helpers into separate submodule
